handling payemnt,add corse to the cart,remove it from the cart

+++ src/main/java/api/dev/students/dto/PayementDto.java
@@ -1,15 +1,16 @@
 package api.dev.students.dto;
 
+import java.math.BigDecimal;
 
 public class PayementDto {
     private String courseName;
-    private long amount;
+    private BigDecimal amount;
 
     
     public String getCourseName() {
         return courseName;
     }
-    public long getAmount() {
+    public BigDecimal getAmount() {
         return amount;
     }
